move editors and layouts to core
